% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{cv_pdnr}
\alias{cv_pdnr}
\title{Variable selection of PDNR using cross validation}
\usage{
cv_pdnr(
  x,
  y,
  hs,
  ns,
  ms,
  taus,
  expand = TRUE,
  eval = NULL,
  k = 5,
  kernel = "epan",
  shape = "general",
  method = "reflection",
  dist_init = "exponential",
  min_iter = 20,
  max_iter = 400,
  tol = 0.001,
  check = 5,
  ncores = NULL
)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{hs}{Possible values of h}

\item{ns}{Possible values of n}

\item{ms}{Possible values of m}

\item{taus}{Possible values of tau}

\item{expand}{Logical: Use of expand.grid or not for inputs?}

\item{eval}{Evaluation points}

\item{k}{Number of folds}

\item{kernel}{Type of kernel: epan, gaussian, or uniform}

\item{shape}{Shape of estimator: general or monotone}

\item{method}{kernel density estimation/sampling method: reflection or transformation}

\item{dist_init}{Distribution for initialization}

\item{min_iter}{Minimum number of iteration}

\item{max_iter}{Maximum number of iteration}

\item{tol}{Error tolerance}

\item{check}{Iteration tolerance}

\item{ncores}{Number of cores}
}
\value{
List
}
\description{
Variable selection of PDNR using cross validation
}
\examples{
\dontrun{
x = seq(0.1, 10, by = 0.1)
y = ftn_wave(x, 0, 1, 1)
y = 1 - y
y_err = y + rnorm(length(y), 0, 0.2)
hs = c(0.1, 0.2)
ns = c(100, 200)
ms = c(10, 20)
taus = c(0.1, 0.2)
expand = TRUE
eval = NULL
k = 5
kernel = "epan"
shape = "general"
method = "reflection"
dist_init = "exponential"
min_iter = 20
max_iter = 400
tol = 1e-3
check = 5
ncores = NULL

ptm <- proc.time()
output <- cv_pdnr(x = x, y = y_err, hs = hs, ns = ns, ms = ms, taus = taus, expand = expand, eval = eval, k = k, kernel = kernel, shape = shape, method = method, dist_init = dist_init, min_iter = min_iter, max_iter = max_iter, tol = tol, check = check, ncores = ncores)
ptm <- proc.time() - ptm # ~ 1.5 hr
}
}
