% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{eval_sse}
\alias{eval_sse}
\title{Computing SSE of an estimator obtained by the IDEA algorithm}
\usage{
eval_sse(storage, x, y, h, kernel = "epan", shape = "general")
}
\arguments{
\item{storage}{data storage}

\item{x}{Input (e.g., distance)}

\item{y}{Output (e.g., correlation value)}

\item{h}{Bandwidth}

\item{kernel}{Type of kernel: epan, gaussian, or uniform}

\item{shape}{Shape of estimator: general or monotone}
}
\value{
SSE
}
\description{
Computing SSE of an estimator obtained by the IDEA algorithm
}
\examples{
# No example
}
