% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{ftn_exponential}
\alias{ftn_exponential}
\title{Exponential isotropic semivariogram model}
\usage{
ftn_exponential(h, c_0, c_1, a)
}
\arguments{
\item{h}{h}

\item{c_0}{c0}

\item{c_1}{c1}

\item{a}{a}
}
\value{
Numeric value
}
\description{
Exponential isotropic semivariogram model
}
\examples{
x = seq(0, 15, by = 0.001)
y = ftn_exponential(x, 0, 1, 1)
plot(x, y, type="l")
}
