% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{ftn_linear}
\alias{ftn_linear}
\title{Linear isotropic semivariogram model}
\usage{
ftn_linear(h, c, b)
}
\arguments{
\item{h}{h}

\item{c}{c}

\item{b}{b}
}
\value{
Numeric value
}
\description{
Linear isotropic semivariogram model
}
\examples{
x = seq(0, 15, by = 0.001)
y = ftn_linear(x, 0, 1)
plot(x, y, type="l")
}
