% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{ftn_power}
\alias{ftn_power}
\title{Power isotropic semivariogram model}
\usage{
ftn_power(h, c, b, lambda)
}
\arguments{
\item{h}{h}

\item{c}{c}

\item{b}{b}

\item{lambda}{lambda}
}
\value{
Numeric value
}
\description{
Power isotropic semivariogram model
}
\examples{
x = seq(0, 15, by = 0.001)
y = ftn_power(x, 0, 1, 1)
plot(x, y, type="l")
}
