% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idea.R
\name{IDEA}
\alias{IDEA}
\title{Iterated Density Estimation Evolutionary Algorithm (IDEA) for regression}
\usage{
IDEA(
  x,
  y,
  h,
  n,
  m,
  tau,
  kernel = "epan",
  shape = "general",
  method = "reflection",
  dist_init = "exponential",
  min_iter = 20,
  max_iter = 400,
  tol = 0.001,
  check = 5
)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{h}{Bandwidth}

\item{n}{n}

\item{m}{m}

\item{tau}{tau}

\item{kernel}{Type of kernel: epan, gaussian, or uniform}

\item{shape}{Shape of estimator: general or monotone}

\item{method}{kernel density estimation/sampling method: reflection or transformation}

\item{dist_init}{Distribution for initialization}

\item{min_iter}{Minimum number of iteration}

\item{max_iter}{Maximum number of iteration}

\item{tol}{Error tolerance}

\item{check}{Iteration tolerance}
}
\value{
List
}
\description{
Iterated Density Estimation Evolutionary Algorithm (IDEA) for regression
}
\examples{
# No example
}
