% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{krnl}
\alias{krnl}
\title{Kernel density estimation using pseudo data}
\usage{
krnl(x, h, data, kernel = "epan", method = "reflection")
}
\arguments{
\item{x}{x}

\item{h}{Bandwidth}

\item{data}{pseudo data}

\item{kernel}{Type of kernel: epan, gaussian, or uniform}

\item{method}{Method of sampling: transformation or reflection}
}
\value{
Values of kernel density estimation
}
\description{
Kernel density estimation using pseudo data
}
\examples{
data <- rexp(1000)
x <- seq(min(data), 5, by = 0.05)

par(mfrow = c(1, 2))
y <- krnl(x, 0.01, data, "epan", "reflection")
z <- krnl(x, 0.01, data, "epan", "transformation")
plot(x,y, type="l", ylim=c(0,2), col="blue")
lines(x,z, col="red")
lines(x, dexp(x))

y <- krnl(x, 1, data, "epan", "reflection")
z <- krnl(x, 1, data, "epan", "transformation")
plot(x,y, type="l", ylim=c(0,2), col="blue")
lines(x,z, col="red")
lines(x, dexp(x))
par(mfrow = c(1, 1))
}
