% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{ksreflect}
\alias{ksreflect}
\title{Kernel-type estimator based on the reflection method}
\usage{
ksreflect(r, data, h, kernel = "epan", shape = "general")
}
\arguments{
\item{r}{Distance}

\item{data}{Pseudo data}

\item{h}{Bandwidth}

\item{kernel}{Type of kernel: epan, gaussian, or uniform}

\item{shape}{Shape of estimator: general or monotone}
}
\value{
Numeric value of the estimator
}
\description{
Kernel-type estimator based on the reflection method
}
\examples{
r <- seq(0.05, 5, by = 0.05)
data <- rexp(10)
v1 <- c() ; v2 <- c() ; v3 <- c()
v4 <- c() ; v5 <- c() ; v6 <- c()

for(i in 1:length(r)) v1[i] <- ksreflect(r = r[i], data = data, h = 1, kernel = "epan", shape = "general")
for(i in 1:length(r)) v2[i] <- ksreflect(r = r[i], data = data, h = 1, kernel = "gaussian", shape = "general")
for(i in 1:length(r)) v3[i] <- ksreflect(r = r[i], data = data, h = 1, kernel = "uniform", shape = "general")
for(i in 1:length(r)) v4[i] <- ksreflect(r = r[i], data = data, h = 1, kernel = "epan", shape = "monotone")
for(i in 1:length(r)) v5[i] <- ksreflect(r = r[i], data = data, h = 1, kernel = "gaussian", shape = "monotone")
for(i in 1:length(r)) v6[i] <- ksreflect(r = r[i], data = data, h = 1, kernel = "uniform", shape = "monotone")

par(mfrow = c(1, 2))
plot(r, v1, type="l", col = "red", ylim = c(-0.5, 2))
lines(r, v2, col="blue")
lines(r, v3, col="green")

plot(r, v4, type="l", col = "red", ylim = c(-0.5, 2))
lines(r, v5, col="blue")
lines(r, v6, col="green")
par(mfrow = c(1, 1))
}
