% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ksreflect_C}
\alias{ksreflect_C}
\title{Kernel-type estimator based on the reflection method (C++ version)}
\usage{
ksreflect_C(r, data, h, kernel, shape)
}
\arguments{
\item{r}{Distance}

\item{data}{Pseudo data}

\item{h}{Bandwidth}

\item{kernel}{Type of kernel: epan, gaussian, or uniform}

\item{shape}{Shape of estimator: general or monotone}
}
\value{
Numeric value of the estimator
}
\description{
Kernel-type estimator based on the reflection method (C++ version)
}
