% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{mse_pdnr}
\alias{mse_pdnr}
\title{Computing MS(P)E of PDNR model}
\usage{
mse_pdnr(
  x_train,
  y_train,
  x_test,
  y_test,
  h,
  n,
  m,
  tau,
  kernel = "epan",
  shape = "general",
  method = "reflection",
  dist_init = "exponential",
  min_iter = 20,
  max_iter = 400,
  tol = 0.001,
  check = 5
)
}
\arguments{
\item{x_train}{x_train}

\item{y_train}{y_train}

\item{x_test}{x_test}

\item{y_test}{y_test}

\item{h}{Bandwidth}

\item{n}{n}

\item{m}{m}

\item{tau}{tau}

\item{kernel}{Type of kernel: epan, gaussian, or uniform}

\item{shape}{Shape of estimator: general or monotone}

\item{method}{kernel density estimation/sampling method: reflection or transformation}

\item{dist_init}{Distribution for initialization}

\item{min_iter}{Minimum number of iteration}

\item{max_iter}{Maximum number of iteration}

\item{tol}{Error tolerance}

\item{check}{Iteration tolerance}
}
\value{
MS(P)E
}
\description{
Computing MS(P)E of PDNR model
}
\examples{
# No example
}
