% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{pdnr}
\alias{pdnr}
\title{Positive defnite nonparametric regression (PDNR)}
\usage{
pdnr(
  x,
  y,
  h,
  n,
  m,
  tau,
  eval = NULL,
  kernel = "epan",
  shape = "general",
  method = "reflection",
  dist_init = "exponential",
  min_iter = 20,
  max_iter = 400,
  tol = 0.001,
  check = 5
)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{h}{Bandwidth}

\item{n}{n}

\item{m}{m}

\item{tau}{tau}

\item{eval}{Evaluation points}

\item{kernel}{Type of kernel: epan, gaussian, or uniform}

\item{shape}{Shape of estimator: general or monotone}

\item{method}{kernel density estimation/sampling method: reflection or transformation}

\item{dist_init}{Distribution for initialization}

\item{min_iter}{Minimum number of iteration}

\item{max_iter}{Maximum number of iteration}

\item{tol}{Error tolerance}

\item{check}{Iteration tolerance}
}
\value{
List
}
\description{
Positive defnite nonparametric regression (PDNR)
}
\examples{
# No example
}
