% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{rkrnl}
\alias{rkrnl}
\title{Drawing random sample from kernel density estimation}
\usage{
rkrnl(n, data, h, kernel = "epan", method = "reflection")
}
\arguments{
\item{n}{Desired size of random sample}

\item{data}{pseudo data}

\item{h}{Bandwidth}

\item{kernel}{Type of kernel: epan, gaussian, or uniform}

\item{method}{Method of sampling: transformation or reflection}
}
\value{
Vector of the random sample
}
\description{
Drawing random sample from kernel density estimation
}
\examples{
n = 100
data = c(0.5, 10)
h = 5

par(mfrow = c(1, 3))
hist(rkrnl(n, data, h, "epan", "reflection"), main = "Epanechnikov")
hist(rkrnl(n, data, h, "gaussian", "reflection"), main = "Gaussian")
hist(rkrnl(n, data, h, "uniform", "reflection"), main = "Uniform")
par(mfrow = c(1, 1))
}
