% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{pdnr}
\alias{pdnr}
\title{Positive defnite nonparametric regression (PDNR)}
\usage{
pdnr(
  x,
  y,
  h,
  n,
  m,
  tau,
  eval = NULL,
  kernel = "epan",
  shape = "general",
  method = "reflection",
  dist_init = "exponential",
  min_iter = 20,
  max_iter = 400,
  tol = 0.001,
  check = 5
)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{h}{Bandwidth}

\item{n}{n}

\item{m}{m}

\item{tau}{tau}

\item{eval}{Evaluation points}

\item{kernel}{Type of kernel: epan, gaussian, or uniform}

\item{shape}{Shape of estimator: general or monotone}

\item{method}{kernel density estimation/sampling method: reflection or transformation}

\item{dist_init}{Distribution for initialization}

\item{min_iter}{Minimum number of iteration}

\item{max_iter}{Maximum number of iteration}

\item{tol}{Error tolerance}

\item{check}{Iteration tolerance}
}
\value{
List
}
\description{
Positive defnite nonparametric regression (PDNR)
}
\examples{
\dontrun{
x = seq(0.1, 10, by = 0.1)
y = ftn_wave(x, 0, 1, 1)
y = 1 - y
y_err = y + rnorm(length(y), 0, 0.2)

h = 0.08
n = 50
m = 10
tau = 0.2
eval = NULL
kernel = "uniform"
shape = "general"
method = "reflection"
dist_init = "exponential"
min_iter = 20
max_iter = 100
tol = 1e-3
check = 5

ptm <- proc.time()
result <- pdnr(x = x, y = y_err, h = h, n = n, m = m, tau = tau, eval = eval, kernel = kernel, shape = shape, method = method, dist_init = dist_init, min_iter = min_iter, max_iter = max_iter, tol = tol, check = check)
ptm <- proc.time() - ptm # ~30s

plot(x, y_err)
lines(x, y, type="l", col="red")
lines(result$eval, result$yhat)

plot(na.omit(result$history_obj_max), type="l", ylim = c(0, 15))
lines(na.omit(result$history_obj_mean))
lines(na.omit(result$history_obj_max_sel))
lines(na.omit(result$history_obj_min))
}
}
